Map NaNs to upper bound of colormap for GIBS colormaps
Numpy normalization behavior changed and now maps NaNs to upper bound of colormap. This needs to be communicated in colormap CSVs we send to GIBS. So, instead of first entry being 0,0,0 -> NaN, it's now upper bound RGB0 -> NaN. Note that we drop this entry for the padded colormaps that we use to make the imagery.