Force empty images to be assigned upper bound of color map and make transparent
Default behavior of Image.new() is to assign color=0, which is lower bound of color map. We were correctly assigning this as the transparent value. Images containing data have upper bound of color bar declared transparent. So, when we give GIBS a mix of empty and data-ful images, they have inconsistent transparency information. This resulted in Worldview incorrectly rendering pixels opaque that should be transparent. This fix ensures consistency between transparency mapping for empty images and data-ful images.